refactor code

+++ src/main/java/fpt/edu/vn/gms/repository/PartCategoryRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PartCategoryRepository extends JpaRepository<PartCategory, Integer> {
+public interface PartCategoryRepository extends JpaRepository<PartCategory, Long> {
 
 }
